Changes API, cursor improvements

+++ api/src/services/LanguageService.ts
@@ -6,6 +6,11 @@ import { PropertyService } from './PropertyService';
 import { TopicService } from './TopicService';
 import { WordService } from './WordService';
 import { DateTime } from 'luxon';
+import { ChangeService } from './ChangeService';
+import { copy } from 'utils/copy';
+import { ChangeBuilder } from './ChangeBuilder';
+
+const LANGUAGE_DELETION_WORDS_THRESHOLD = 50;
 
 export interface CreateLanguageParams {
   id?: LanguageId;
@@ -28,6 +33,9 @@ export class LanguageService {
     private topicService: TopicService,
     @Inject(forwardRef(() => WordService))
     private wordService: WordService,
+    @Inject(forwardRef(() => ChangeService))
+    private changeService: ChangeService,
+    private changeBuilder: ChangeBuilder,
   ) {}
 
   async getAll(): Promise<Language[]> {
@@ -45,28 +53,49 @@ export class LanguageService {
   async create(params: CreateLanguageParams): Promise<Language> {
     const language: Language = {
       id: params.id ?? uuid(),
-      name: params.name,
-      addedAt: params.addedAt ?? DateTime.now(),
+      name: params.name.trim(),
+      addedAt: params.addedAt ?? DateTime.utc(),
     };
     await this.languageRepository.create(language);
+    await this.changeService.create(
+      this.changeBuilder.buildCreateLanguageChange(language),
+    );
     return language;
   }
 
   async update(params: UpdateLanguageParams): Promise<Language> {
     const language = await this.getById(params.id);
+    const currentLanguage = copy(language);
 
-    language.name = params.name;
-    await this.languageRepository.update(language);
+    language.name = params.name.trim();
+
+    const change = this.changeBuilder.buildUpdateLanguageChange(
+      language,
+      currentLanguage,
+    );
+    if (change) {
+      await this.languageRepository.update(language);
+      await this.changeService.create(change);
+    }
+
     return language;
   }
 
   async delete(id: LanguageId): Promise<Language> {
     const language = await this.getById(id);
 
+    const wordCount = await this.wordService.getCount(language.id);
+    if (wordCount > LANGUAGE_DELETION_WORDS_THRESHOLD) {
+      throw new Error('Language has too many words');
+    }
+
     await this.wordService.deleteForLanguage(id);
     await this.topicService.deleteForLanguage(id);
     await this.propertyService.deleteForLanguage(id);
     await this.languageRepository.delete(id);
+    await this.changeService.create(
+      this.changeBuilder.buildDeleteLanguageChange(language),
+    );
     return language;
   }
 }
